refactoring the code and put it in a folder
