+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/MainActivity.kt
@@ -16,12 +16,11 @@ import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
-import com.example.habittrackerapp.model.Habit
 import com.example.habittrackerapp.layout.MainLayout
 import com.example.habittrackerapp.signInSignUp.rememberMutableStateListOf
 import com.example.habittrackerapp.model.Note
 import com.example.habittrackerapp.ui.theme.HabitTrackerAppTheme
-import com.example.habittrackerapp.model.User
+import com.example.habittrackerapp.model.UserViewModel.User
 import com.google.firebase.FirebaseApp
 
 val LocalNavController = compositionLocalOf<NavController> { error("No NavController found!") }
