+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/auth/AuthRepositoryFirebase.kt
@@ -1,6 +1,6 @@
 package com.example.habittrackerapp.auth
 
-import com.example.habittrackerapp.model.User
+import com.example.habittrackerapp.model.UserViewModel.User
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
 import kotlinx.coroutines.flow.MutableStateFlow
