+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/UserViewModel/SavedUserViewModel.kt
@@ -1,9 +1,10 @@
-package com.example.habittrackerapp.model
+package com.example.habittrackerapp.model.UserViewModel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.example.habittrackerapp.MyApp
+import com.example.habittrackerapp.model.SavedUser
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
