+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/UserViewModel/User.kt
@@ -1,4 +1,4 @@
-package com.example.habittrackerapp.model
+package com.example.habittrackerapp.model.UserViewModel
 
 import android.os.Parcelable
 import kotlinx.coroutines.flow.StateFlow
