+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/UserViewModel/UserRepositoryDatastore.kt
@@ -1,12 +1,12 @@
-package com.example.habittrackerapp.model
+package com.example.habittrackerapp.model.UserViewModel
 
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.intPreferencesKey
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
+import com.example.habittrackerapp.model.SavedUser
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
