+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/settings/DeleteUser.kt
@@ -18,10 +18,10 @@ import com.example.habittrackerapp.auth.AuthViewModelFactory
 import com.example.habittrackerapp.data
 import com.example.habittrackerapp.model.NotesViewModel
 import com.example.habittrackerapp.model.NotesViewModelFactory
-import com.example.habittrackerapp.model.SavedUserViewModel
-import com.example.habittrackerapp.model.SavedUserViewModelSavedFactory
-import com.example.habittrackerapp.model.UserViewModel
-import com.example.habittrackerapp.model.UserViewModelFactory
+import com.example.habittrackerapp.model.UserViewModel.SavedUserViewModel
+import com.example.habittrackerapp.model.UserViewModel.SavedUserViewModelSavedFactory
+import com.example.habittrackerapp.model.UserViewModel.UserViewModel
+import com.example.habittrackerapp.model.UserViewModel.UserViewModelFactory
 
 @Composable
 fun DeleteUser(MyViewModel: UserViewModel = viewModel(factory= UserViewModelFactory()),
