+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/signInSignUp/SignInSignUpScreen.kt
@@ -15,10 +15,10 @@ import com.example.habittrackerapp.LocalNavController
 import com.example.habittrackerapp.auth.AuthViewModel
 import com.example.habittrackerapp.auth.AuthViewModelFactory
 import com.example.habittrackerapp.data
-import com.example.habittrackerapp.model.SavedUserViewModel
-import com.example.habittrackerapp.model.SavedUserViewModelSavedFactory
-import com.example.habittrackerapp.model.UserViewModel
-import com.example.habittrackerapp.model.UserViewModelFactory
+import com.example.habittrackerapp.model.UserViewModel.SavedUserViewModel
+import com.example.habittrackerapp.model.UserViewModel.SavedUserViewModelSavedFactory
+import com.example.habittrackerapp.model.UserViewModel.UserViewModel
+import com.example.habittrackerapp.model.UserViewModel.UserViewModelFactory
 
 @Composable
 fun SignSignUpScreen(savedUserViewModel: SavedUserViewModel = viewModel(factory = SavedUserViewModelSavedFactory()),
