+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/signInSignUp/UserSignUpScreen.kt
@@ -43,11 +43,10 @@ import com.example.habittrackerapp.auth.AuthViewModelFactory
 import com.example.habittrackerapp.LocalNavController
 import com.example.habittrackerapp.R
 import com.example.habittrackerapp.data
-import com.example.habittrackerapp.model.SavedUserViewModel
-import com.example.habittrackerapp.model.SavedUserViewModelSavedFactory
-import com.example.habittrackerapp.model.User
-import com.example.habittrackerapp.model.UserViewModel
-import com.example.habittrackerapp.model.UserViewModelFactory
+import com.example.habittrackerapp.model.UserViewModel.SavedUserViewModel
+import com.example.habittrackerapp.model.UserViewModel.SavedUserViewModelSavedFactory
+import com.example.habittrackerapp.model.UserViewModel.UserViewModel
+import com.example.habittrackerapp.model.UserViewModel.UserViewModelFactory
 import com.example.habittrackerapp.ui.theme.HabitTrackerAppTheme
